% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsor.R
\name{Collapsor}
\alias{Collapsor}
\title{collapse a single lineage}
\usage{
Collapsor(file = NULL, url = NULL)
}
\arguments{
\item{compressed_lineage}{GISAID credentials.}

\item{potential_parents}{parental lineages - get this from XXXX}

\item{strict}{strictly match collapsed definition? will return NA if not matching}
}
\value{
str of collapsed lineage
}
\description{
collapse a single lineage
}
